BUGFIX subscription lists error without api keys
After a base install the Constant Contact tab throws an error from trying to create a MultiValueDropdownField with no data. After saving the API KEY and ACCESS TOKEN the error goes away as the module has access to the Constant Contact data.

This fix checks if the API KEY and ACCESS TOKEN are saved and if not will show a message rather than attempting the dropdown.